Restructured to make more consistent and node friendly
This removes all bower stuff, the grunt crap, the checked in
dependencies etc. (Although, to be fair, I'm actually a
proponent of checking in dependencies; more on that another
time.)

All tests now run under node mocha, but not in a browser
unfortunately because of the need for a build step. I'll look
into that in subsequent work.