Remove domino dependency and prepare for npm publishing
This removes the domino dependency because it was huge, and can
be replaced with something smarter since we just need to shim
matchesSelector (I think.) See #21 for some further discussion.

All tests should be sorted, and the Makefile will build and run
the tests provided html-files are handled by a browser. These
will probably change to a TAP harness in the future, just like I
changed rhumb.