Better tests; got rid of explicit browser dependencies
There are still browser dependencies, so this doesn't sort out
node, but further work is necessary to make this run in a browser.

Tests are now producing TAP output, so the mocha dependency is
gone. Also chai and sinon are gone. The tests were also split
in multiple files to make things a bit more reasonable.